requirements.txt
cffi==1.15.1
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.3.0
pycparser==2.21
pyvirtualcam==0.10.2
